### 2020-07-22 #390 @KevinShook
CoSMoS - does our package fit?
Submitting Author: Kevin Shook@https://github.com/KevinShook
Repository:   https://github.com/TycheLab/CoSMoS

---

```
Package: CoSMoS
Title: Complete Stochastic Modelling Solution
Version: 2.0.0
Authors@R: c(person("Filip", "Strnad",
                     role = "aut"),
             person("Simon Michael", "Papalexiou",
                      role = c("aut")),
             person("Francesco", "Serinaldi",
                      role = "aut"),
             person("Yannis", "Markonis",
                      role = "aut"),
             person("Kevin", "Shook",
                      role = c("ctb", "cre"),
                      email = "kevin.shook@usask.ca")
              )
Description: A single framework, unifying, extending, and improving a general-purpose modelling strategy, based on the assumption that any process can emerge by transforming a specific 'parent' Gaussian process Papalexiou (2018) <doi:10.1016/j.advwatres.2018.02.013>.
Depends: R (>= 3.5.0), ggplot2, data.table
    grDevices,
    nloptr,
    MBA,
    Matrix,
    mAr,
    matrixcalc,
    mvtnorm,
    plotly,
    cowplot,
    directlabels,
    animation,
    pracma
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

	-  geospatial data

The package would appear to fit under the geospatial data category as it generates synthetic geospatial data.
I see that the **charlatan** package on rOpenSci also generates fake data, although of completely different kinds - perhaps this is also a model for CoSMoS.

The main users of CoSMoS are scientists or practitioners who want to generate synthetic environmental data. Synthetic data are useful for designing water-resource infrastructure, or for forcing physical models, as historical time series are often too short to be useful. Synthetic data also allow the creation of many realisations of data sets having the same properties, which allows the determination of the sensitivity of a model to its input forcings.

There are no other packages on CRAN or rOpenSci which do the same things as CoSMoS.

NA. No human data is involved.
The package is currently on CRAN: https://cran.r-project.org/package=CoSMoS
