### 2020-11-03 #390 @melvidoni
Hello @KevinShook. Apologies for the delay, it was extremely long. We have decided that this package is in-scope giving the geospatial component, so we welcome a full submission. Please, feel free to submit whenever is possible for you.
